+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/ChannelRepository.kt
@@ -8,6 +8,7 @@ import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.livedata.dao.ChannelDao
 import io.getstream.chat.android.livedata.entity.ChannelEntity
 import io.getstream.chat.android.livedata.extensions.isPermanent
+import io.getstream.chat.android.livedata.repository.mapper.toEntity
 
 internal class ChannelRepository(
     var channelDao: ChannelDao,
