+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/repository/ReactionRepositoryTest.kt
@@ -4,6 +4,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.common.truth.Truth
 import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.livedata.BaseDomainTest
+import io.getstream.chat.android.livedata.repository.mapper.toModel
 import kotlinx.coroutines.runBlocking
 import org.junit.Test
 import org.junit.runner.RunWith
